+++ packages/blip/app/components/chart/patientDashboard.js
@@ -3,12 +3,13 @@ import PropTypes from 'prop-types'
 import moment from 'moment-timezone'
 import i18next from 'i18next'
 import Box from '@mui/material/Box'
-import PatientStatistics from './patientStatistics'
 import DeviceUsage from './deviceUsage'
 import './patientDashboardVars.css'
 import AccessTime from '@mui/icons-material/AccessTime'
 import RemoteMonitoringWidget from 'yourloops/components/dashboard-widgets/remote-monitoring-widget'
 import { useTeam } from 'yourloops/lib/team'
+import { PatientStatisticsWidget } from 'yourloops/components/dashboard-widgets/patient-statistics-widget'
+import Stats from './stats'
 
 const t = i18next.t.bind(i18next)
 
@@ -27,13 +28,13 @@ const PatientDashboard = (props) => {
     alarmCard: AlarmCard,
     medicalFilesWidget: MedicalFilesWidget,
     //eslint-disable-next-line
-    timePrefs, tidelineData, trackMetric, onSwitchToDaily, userIsHCP, isSelectedTeamMedical
+    timePrefs, medicalDataService, trackMetric, onSwitchToDaily, userIsHCP, isSelectedTeamMedical
   } = props
   const isMonitoringEnabled = patient.monitoring?.enabled
   const shouldDisplayChatWidget = isMonitoringEnabled && (!userIsHCP || isSelectedTeamMedical)
 
   const { getMedicalTeams } = useTeam()
-
+  const { medicalData } = medicalDataService
   const showRemoteMonitoringWidget = !user.isUserCaregiver() && getMedicalTeams().some(team => team.monitoring?.enabled)
 
   const getEndpoints = () => {
@@ -58,21 +59,27 @@ const PatientDashboard = (props) => {
             userRole={user.role}
           />
         }
-        <PatientStatistics
-          bgPrefs={bgPrefs}
-          bgSource={dataUtil.bgSource}
-          chartPrefs={chartPrefs}
-          chartType="patientStatistics"
+        <PatientStatisticsWidget
           dataUtil={dataUtil}
+          bgPrefs={bgPrefs}
           endpoints={endpoints}
-          loading={loading}
-          parametersConfig={tidelineData.medicalData?.pumpSettings[0]?.payload?.parameters}
-        />
+        >
+          <Stats
+            bgPrefs={bgPrefs}
+            bgSource={dataUtil.bgSource}
+            chartPrefs={chartPrefs}
+            chartType="patientStatistics"
+            dataUtil={dataUtil}
+            endpoints={endpoints}
+            loading={loading}
+            parametersConfig={medicalData?.pumpSettings[0]?.payload?.parameters}
+          />
+        </PatientStatisticsWidget>
         <DeviceUsage
           bgPrefs={bgPrefs}
           timePrefs={timePrefs}
           patient={patient}
-          tidelineData={tidelineData}
+          tidelineData={medicalDataService}
           trackMetric={trackMetric}
           dataUtil={dataUtil}
           chartPrefs={chartPrefs}
@@ -110,7 +117,7 @@ PatientDashboard.propTypes = {
   msRange: PropTypes.number.isRequired,
   onSwitchToDaily: PropTypes.func.isRequired,
   canPrint: PropTypes.bool,
-  tidelineData: PropTypes.object.isRequired
+  medicalDataService: PropTypes.object.isRequired
 }
 
 export default PatientDashboard