+++ packages/dumb/src/index.ts
@@ -60,6 +60,7 @@ import { type BgPrefs } from './models/blood-glucose.model'
 import { RangeSelect } from './components/trends/cbg/range-select/range-select'
 import { TrendsProvider } from './provider/trends.provider'
 import { FocusedRangeLabels } from './components/trends/common/focused-range-labels/focused-range-labels'
+import { UnitsType } from './models/enums/units-type.enum'
 import {
   TrendsSvgContainerSized as TrendsSvgContainer
 } from './components/trends/common/trends-svg-container/trends-svg-container'
@@ -68,6 +69,7 @@ import { FocusedCbgSliceSegmentMemoized as FocusedCbgSliceSegment } from './comp
 export {
   AverageDailyDoseStat,
   type BgPrefs,
+  UnitsType,
   BloodGlucoseTooltip,
   BolusTooltip,
   CbgDateTraceLabel,